Add control for toggling color spectrum range
The color spectrum in Minecraft map-art is increased by 3x when varying
the height of placed blocks. This does however make placing the blocks
in survival significantly harder.

The full color spectrum was enabled by default, this change makes it
toggleable in case people want an easier map to build at the cost of
fewer colors.