Refactor palette data and data structures
There were a lot of mixed types that were all slightly similar in nature
which was making following the flow of data difficult, especially when
making changes. This is a clean up and simplification of the data
formats in use, and adds support for some additional features.

The block palette JSON data was moved out to its own package and was
updated to support multiple MC versions. The data was also augmented
with missing nbt property data such as the axis on which to place blocks
in order to achieve the denoted color.

The pixel->block grid conversion function was updated to support
balancing over multiple selected blocks, though support for this hasn't
yet been added to the interface. This will come in follow up commits.

Color definitions were normalized into the Pixel type instead of some
being Pixel and others RGBTuple.