Improve image processing pipeline performance
This drastically improves the image processing pipelines performance by
significantly reducing the number of iterations performed on pixels and
by removing a lot of the intermediary translation phases.

Up and down scaling is now performed at the same time as pixel
transformation - which itself is structured as a per-pixel transformer
pipeline. This means we only need to iterate over pixel data twice. Once
to down-scale and once to upscale.